postbox/requirements.txt
flask
redis
pyopenssl
flask-jwt-extended
bcrypt
Flask-Session
python-dotenv